Restructure actions builtins
